+++ software/distro/setup/base-os/forklift/precache-image.sh
@@ -23,7 +23,7 @@ echo "Downloading $image to $precached_image..."
 if ! $crane --platform "$platform" pull --format legacy "$image" "$precached_image"
 then
   echo "Encountered error, trying one more time to download $image..."
-  rm -f "$precache_image" || sudo rm -f "$precached_image"
+  rm -f "$precached_image" || sudo rm -f "$precached_image"
   $crane --platform "$platform" pull --format legacy "$image" "$precached_image"
 fi
 echo "Finished downloading $image!"